+++ client/packages/common/src/plugins/components/PluginProvider.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   ColumnPlugin,
   ComponentPlugin,
@@ -107,20 +106,3 @@ export const usePluginProvider = create<PluginProvider>((set, get) => ({
       ),
     })),
 }));
-
-export const usePluginEvents = () => {
-  const addEventListener = usePluginProvider(state => state.addEventListener);
-  const removeEventListener = usePluginProvider(
-    state => state.removeEventListener
-  );
-  const dispatchEvent = usePluginProvider(state => state.dispatchEvent);
-
-  return React.useMemo(
-    () => ({
-      addEventListener,
-      dispatchEvent,
-      removeEventListener,
-    }),
-    [addEventListener, dispatchEvent, removeEventListener]
-  );
-};
