+++ client/packages/common/src/plugins/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './usePluginColumns';
 export * from './usePluginElements';
+export * from './usePluginEvents';
